requirements.txt
python-frontmatter==1.1.0
pyyaml==6.0.1
python-dotenv==1.0.0
requests==2.31.0
google-generativeai==0.3.0
